requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
yt-dlp
python-dotenv
supabase
google-genai
